Refactor source_provider to its own module
